refactor: change on_policy structure
