+++ omnisafe/algorithms/on_policy/naive_lagrange/trpo_lag.py
@@ -12,23 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Implementation of the CPPO Pid-Lagrange algorithm."""
+"""Implementation of the Lagrange version of TRPO algorithm."""
 
 import torch
 
 from omnisafe.algorithms import registry
-from omnisafe.algorithms.on_policy.policy_gradient import PolicyGradient
-from omnisafe.common.pid_lagrange import PIDLagrangian
+from omnisafe.algorithms.on_policy.base.trpo import TRPO
+from omnisafe.common.lagrange import Lagrange
 
 
 @registry.register
-class CPPOPid(PolicyGradient, PIDLagrangian):
-    """The Responsive Safety in Reinforcement Learning by PID Lagrangian Methods.
+class TRPOLag(TRPO, Lagrange):
+    """The Lagrange version of TRPO algorithm.
 
     References:
-        Paper Name: Responsive Safety in Reinforcement Learning by PID Lagrangian Methods.
-        Paper author: Joshua Achiam, David Held, Aviv Tamar, Pieter Abbeel.
-        Paper URL: https://arxiv.org/abs/1705.10528
+        Paper Name: Benchmarking Safe Exploration in Deep Reinforcement Learning.
+        Paper author: Alex Ray, Joshua Achiam, Dario Amodei
+        Paper URL: https://cdn.openai.com/safexp-short.pdf
 
     """
 
@@ -36,43 +36,40 @@ class CPPOPid(PolicyGradient, PIDLagrangian):
         self,
         env_id,
         cfgs,
-        algo: str = 'CPPO-PID',
+        algo: str = 'TRPO-Lag',
         wrapper_type: str = 'OnPolicyEnvWrapper',
     ):
-
-        PolicyGradient.__init__(
+        """initialize"""
+        TRPO.__init__(
             self,
             env_id=env_id,
             cfgs=cfgs,
             algo=algo,
             wrapper_type=wrapper_type,
         )
-        PIDLagrangian.__init__(self, **self.cfgs.PID_cfgs._asdict())
-
-        self.clip = self.cfgs.clip
-        self.cost_limit = self.cfgs.cost_limit
+        Lagrange.__init__(
+            self,
+            cost_limit=self.cfgs.lagrange_cfgs.cost_limit,
+            lagrangian_multiplier_init=self.cfgs.lagrange_cfgs.lagrangian_multiplier_init,
+            lambda_lr=self.cfgs.lagrange_cfgs.lambda_lr,
+            lambda_optimizer=self.cfgs.lagrange_cfgs.lambda_optimizer,
+        )
 
     def algorithm_specific_logs(self):
         super().algorithm_specific_logs()
-        self.logger.log_tabular('Metrics/LagrangeMultiplier', self.cost_penalty)
-        self.logger.log_tabular('PID/pid_Kp', self.pid_kp)
-        self.logger.log_tabular('PID/pid_Ki', self.pid_ki)
-        self.logger.log_tabular('PID/pid_Kd', self.pid_kd)
+        self.logger.log_tabular('Metrics/LagrangeMultiplier', self.lagrangian_multiplier.item())
 
-    def compute_loss_pi(self, data: dict):
-        """compute loss for policy"""
+    def compute_loss_pi(self, data: dict) -> tuple:
+        # Policy loss
         dist, _log_p = self.actor_critic.actor(data['obs'], data['act'])
         ratio = torch.exp(_log_p - data['log_p'])
-        ratio_clip = torch.clamp(ratio, 1 - self.clip, 1 + self.clip)
 
-        surr_adv = (torch.min(ratio * data['adv'], ratio_clip * data['adv'])).mean()
-        surr_cadv = (torch.max(ratio * data['cost_adv'], ratio_clip * data['cost_adv'])).mean()
-
-        loss_pi = -surr_adv
+        loss_pi = -(ratio * data['adv']).mean()
         loss_pi -= self.cfgs.entropy_coef * dist.entropy().mean()
 
-        penalty = self.cost_penalty
-        loss_pi += penalty * surr_cadv
+        # ensure that Lagrange Multiplier is positive
+        penalty = torch.clamp_min(self.lagrangian_multiplier, 0.0)
+        loss_pi += penalty * (ratio * data['cost_adv']).mean()
         loss_pi /= 1 + penalty
 
         # Useful extra info
@@ -83,12 +80,15 @@ class CPPOPid(PolicyGradient, PIDLagrangian):
         return loss_pi, pi_info
 
     def update(self):
-        """update policy"""
+        """Update."""
+        # pre-process data
         raw_data, data = self.buf.pre_process_data()
+        # sub-sampling accelerates calculations
+        self.fvp_obs = data['obs'][::4]
         # Note that logger already uses MPI statistics across all processes..
         ep_costs = self.logger.get_stats('Metrics/EpCost')[0]
         # First update Lagrange multiplier parameter
-        self.pid_update(ep_costs)
+        self.update_lagrange_multiplier(ep_costs)
         # now update policy and value network
         self.update_policy_net(data=data)
         self.update_value_net(data=data)
